Рефакторинг и кодстайл
